234/requirements.txt
discord.py>=2.3.2
aiosqlite>=0.19.0
asyncio
python-dotenv>=1.0.0
datetime
json5>=0.9.0
typing-extensions>=4.7.0